fix(Zvqwbdota16i): prevent signed int32 overflow in unsigned 16-bit multiply
In vqwbdotau16_vv (altfmt=0, vs1 unsigned), the accumulation:
  dest += int64_t(e1 * e2)
promotes uint16_t e1,e2 to int before multiplying.  When both are
0xFFFF (65535), the product 4294836225 exceeds INT32_MAX (2147483647)
causing undefined signed overflow.  On x86-64 with GCC -O3 the
result wraps to int32_t(-131071), which sign-extends to int64_t
0xFFFFFFFFFFFE0001 instead of the correct 0x00000000FFFE0001.

Fix: cast e1 and e2 to uint32_t before multiplying.  uint32_t can
hold 65535*65535 = 4294836225 without overflow, and the subsequent
cast to int64_t zero-extends the result correctly.

Spec ref (Zvqwbdota16i): SPEC_vs2 elements are unsigned 16-bit for
vqwbdotau; when altfmt=0, SPEC_vs1 elements are also unsigned 16-bit.
The product of two unsigned 16-bit values requires 32 unsigned bits.

+++ dot-product.cpp
@@ -762,7 +762,7 @@ Hart<URV>::vqwbdotau16_vv(const DecodedInst* di, unsigned s1gx8, unsigned s2gx8,
               if (op2Signed)
                 dest += int64_t(e1 * std::bit_cast<int16_t>(e2));
               else
-                dest += int64_t(e1 * e2);
+                dest += int64_t(uint32_t(e1) * uint32_t(e2));  // uint32 avoids signed overflow when both are 65535
           }
         }
       else if (vecRegs_.isMaskAgnostic() and vecRegs_.isMaskAgnosticOnes())
